requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
python-dotenv
firebase-admin
